+++ .github/workflows/build.yml
@@ -25,7 +25,7 @@ jobs:
   phpunit:
     uses: yiisoft/actions/.github/workflows/phpunit.yml@master
     with:
-      extensions: intl
+      extensions: :intl
       coverage: xdebug
       os: >-
         ['ubuntu-latest', 'windows-latest']
@@ -40,3 +40,13 @@ jobs:
         ['ubuntu-latest', 'windows-latest']
       php: >-
         ['8.1']
+  phpunit-with-intl:
+    uses: yiisoft/actions/.github/workflows/phpunit.yml@master
+    with:
+      phpunitConfig: phpunit-with-intl.xml
+      extensions: intl
+      coverage: xdebug
+      os: >-
+        ['ubuntu-latest', 'windows-latest']
+      php: >-
+        ['8.0', '8.1']
