+++ tests/ValidatorTest.php
@@ -1297,26 +1297,12 @@ class ValidatorTest extends TestCase
         );
     }
 
-    public function testDefaultTranslatorWithIntl(): void
-    {
-        if (!extension_loaded('intl')) {
-            $this->markTestSkipped('The intl extension must be available for this test.');
-        }
-
-        $this->checkDefaultTranslator('3-few');
-    }
-
     public function testDefaultTranslatorWithoutIntl(): void
     {
         if (extension_loaded('intl')) {
             $this->markTestSkipped('The intl extension must be unavailable for this test.');
         }
 
-        $this->checkDefaultTranslator('3');
-    }
-
-    private function checkDefaultTranslator(string $expectedErrorMessage): void
-    {
         $data = ['number' => 3];
         $rules = [
             'number' => new Number(
@@ -1325,9 +1311,9 @@ class ValidatorTest extends TestCase
                 tooBigMessage: '{value, selectordinal, one{#-one} two{#-two} few{#-few} other{#-other}}',
             ),
         ];
-        $validator = new Validator(new SimpleRuleHandlerContainer());
+        $validator = new Validator();
 
         $result = $validator->validate($data, $rules);
-        $this->assertSame(['number' => [$expectedErrorMessage]], $result->getErrorMessagesIndexedByPath());
+        $this->assertSame(['number' => ['3']], $result->getErrorMessagesIndexedByPath());
     }
 }
